backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
pandas
xgboost
python-dotenv
psycopg2-binary
scikit-learn
matplotlib
fpdf
